docs(examples): consolidate AI examples into a single using-ai
