+++ docs/examples/examples.json
@@ -16,9 +16,5 @@
   { "title": "File Binding", "description": "Bind uploaded files directly to structs for type-safe file processing.", "tags": ["Files", "Data Binding"], "href": "https://github.com/gofr-dev/gofr/tree/main/examples/using-file-bind" },
   { "title": "HTML Templates", "description": "Serve server-rendered HTML pages with Go template support.", "tags": ["Templates", "Frontend"], "href": "https://github.com/gofr-dev/gofr/tree/main/examples/using-html-template" },
   { "title": "CLI Application", "description": "Build command-line tools with subcommands, flags, and structured output.", "tags": ["CLI", "Tools"], "href": "https://github.com/gofr-dev/gofr/tree/main/examples/sample-cmd" },
-  { "title": "Calling an LLM", "description": "Call a provider-agnostic, OpenAI-compatible LLM from a handler with automatic tracing, token metrics and health checks.", "tags": ["AI", "LLM"], "href": "https://github.com/gofr-dev/gofr/tree/main/examples/using-ai", "featured": true },
-  { "title": "Streaming LLM Responses", "description": "Stream model tokens to the client as server-sent events with response.Stream.", "tags": ["AI", "LLM", "Streaming"], "href": "https://github.com/gofr-dev/gofr/tree/main/examples/using-ai-stream" },
-  { "title": "MCP Tools", "description": "Expose your handlers to AI agents as Model Context Protocol tools with app.EnableMCP.", "tags": ["AI", "MCP", "Agents"], "href": "https://github.com/gofr-dev/gofr/tree/main/examples/using-mcp" },
-  { "title": "Agent Loop", "description": "Drive a userland agent loop over your service's own handlers with ctx.LLM().Tools().", "tags": ["AI", "Agents"], "href": "https://github.com/gofr-dev/gofr/tree/main/examples/using-agent" },
-  { "title": "AI, Streaming, MCP and Agent", "description": "One service showing the whole GoFr AI surface: LLM calls, streaming, MCP tools and an agent loop.", "tags": ["AI", "LLM", "MCP", "Agents"], "href": "https://github.com/gofr-dev/gofr/tree/main/examples/using-ai-agent", "featured": true }
+  { "title": "AI: LLM, Streaming, MCP and Agents", "description": "One service showing the whole GoFr AI surface: LLM calls, token-streaming, exposing handlers as MCP tools, and a userland agent loop over ctx.LLM().Tools().", "tags": ["AI", "LLM", "MCP", "Agents"], "href": "https://github.com/gofr-dev/gofr/tree/main/examples/using-ai", "featured": true }
 ]
